+++ src/Eve-O-Preview/Configuration/Interface/IThumbnailConfiguration.cs
@@ -22,6 +22,7 @@ namespace EveOPreview.Configuration
 		bool EnableClientLayoutTracking { get; set; }
 		bool HideActiveClientThumbnail { get; set; }
 		bool ShowThumbnailPreviews { get; set; }
+		bool PreviewMinimizedClients { get; set; }
 		bool HideLoginClientThumbnail { get; set; }
 		bool MinimizeInactiveClients { get; set; }
 		bool HideCaptionOnClients { get; set; }